requirements.txt
streamlit 
pandas 
google-generativeai
